Add 1st class support for pixi hooks
This hook will install the `pixi` tool into the hook's environment if
it's missing. If a `pixi.toml` exists in the prefix it will be used as
the environment configuration, additional_dependencies are supported to
extend the environment.

The run command will call Pixi tasks defined in the `pixi.toml`.
Otherwise it will fallback to running executables in the environments
$PATH.

See: https://github.com/prefix-dev/pixi/